Add files via upload
As a Python developer, I successfully completed a project where I implemented web scraping techniques to extract reviews of a product from Flipkart. I used libraries like requests, BeautifulSoup, and pymongo to parse and store the data in both a CSV file and a MongoDB database. I also incorporated error handling to ensure the smooth running of the code. Overall, this project demonstrates my ability to work with web scraping and data storage technologies in Python.